Add other data to new forecast, remove unneeded components.
