+++ web-ui/src/App.css
@@ -34,7 +34,7 @@ body {
 header {
   margin-top: 5px;
   overflow: auto;
-  border-bottom: 1px solid;
+  border-bottom: 1px solid var(--forecast-border);
   border-radius: 0.25em;
   padding: 4px 4px 8px 4px;
 }
@@ -76,7 +76,7 @@ form button:hover, a:hover, .forecastDay button:hover {
 
 form {
   padding: 10px 0;
-  border-bottom: 1px solid var(--black);
+  border-bottom: 1px solid var(--forecast-border);
   margin-bottom: 10px;
 }
 
@@ -240,31 +240,42 @@ th {
   margin: 7px;
 }
 
-.newForecast .precipitation svg {
+.newForecast .otherData svg {
   float: left;
   width: 20px;
   height: 20px;
   margin: 14px 0 0 0;
 }
 
-.newForecast.big .precipitation svg {
+.newForecast.big .otherData svg {
   width: 30px;
   height: 30px;
   margin-left: 10px;
   margin-top: 7px;
 }
 
-.newForecast .precipitation .precipitationPercentage {
+.newForecast .otherData .otherLineOneText {
   float: left;
   padding-top: 21px;
   padding-left: 7px;
   font-size: 0.5em;
 }
 
-.big .precipitation .precipitationPercentage {
+.big .otherData {
   float: left;
+}
+
+.big .otherData .otherLineOneText {
   padding-top: 16px;
-  font-size: 1em;
+  padding-left: 0;
+  font-size: 0.75em;
+}
+
+.big .otherData .otherLineTwoText {
+  float: left;
+  padding-top: 0;
+  padding-left: 12px;
+  font-size: 0.75em;
 }
 
 .newForecast .tempContainer {
@@ -318,7 +329,12 @@ th {
 .big .indicatorContainer svg {
   width: 100px;
   height: 100px;
-  padding-top: 20px;
+}
+
+.big .otherContainer {
+  width: 100%;
+  float: left;
+  margin-bottom: 7px;
 }
 
 .newForecast .windNos {
